DESCRIPCION DE LA TASK
# Chat Translation Microservice

We need a microservice that will handle the real-time translation of chat messages and ensures that we are only requesting translations via the Google API for the phrases in the most cost-efficient way. 

It should also always be as failsafe as possible to guarantee working chat communication - **even when the Google API is offline / not reachable.** 

The technical resources for the microservice will be limited to 4 GB of memory and 2 vCPUs, but it will run on two separate instances concurrently for high availability. 

# Required Features

- Public REST-API Interface
- Loading translations from Google API
- Translation Cache for the most used phrases
- Cache size management
- Failover if Google API is not responding or failing

### Step One (HTTP)

In plain Javascript, start a web server that is listening on port 8080. When visiting [http://localhost:8080/](http://localhost:8080/), it should respond with a JSON message saying “hello world” without any translation logic.

For this task, you need to research a sufficient **library** to use for the web server / accepting the HTTP request.

### Step Two (Arguments)

Find a way how you can translate pre-defined text messages that are passed to the endpoint “/translate”.  The string “hello” should translate to “hola” and “foo” should translate to “bar”. 

 The response should be a JSON message with the original text and the translation. If no translation was found for the input text, an error message should be returned with a reasonable HTTP status code.